refactor: split channel to partial classes
