+++ src/stream-chat-net/APPSettingsWithDetails.cs
@@ -37,9 +37,6 @@ namespace StreamChat
 
         [JsonProperty(NullValueHandling = NullValueHandling.Ignore, PropertyName = "multi_tenant_enabled")]
         public bool MultiTenantEnabled { get; set; }
-
-
-        public AppSettingsWithDetails() { }
     }
 
     public class PushNotificationFields
@@ -49,8 +46,6 @@ namespace StreamChat
 
         [JsonProperty(NullValueHandling = NullValueHandling.Ignore, PropertyName = "firebase")]
         public FirebaseFields Firebase { get; set; }
-
-        public PushNotificationFields() { }
     }
 
     public class APNFields
@@ -74,9 +69,7 @@ namespace StreamChat
         public string TeamID { get; set; }
 
         [JsonProperty(NullValueHandling = NullValueHandling.Ignore, PropertyName = "key_id")]
-        public string KeyID { get; set; }
-
-        public APNFields() { }
+        public string KeyID { get; set; }  
     }
 
     public class FirebaseFields
@@ -86,7 +79,5 @@ namespace StreamChat
 
         [JsonProperty(NullValueHandling = NullValueHandling.Ignore, PropertyName = "enabled")]
         public bool Enabled { get; set; }
-
-        public FirebaseFields() { }
     }
 }
