+++ src/stream-chat-net/Channel.Members.cs
@@ -1,0 +1,58 @@
+using System.Collections.Generic;
+using System.Threading.Tasks;
+using Newtonsoft.Json;
+using Newtonsoft.Json.Linq;
+using StreamChat.Rest;
+
+namespace StreamChat
+{
+    public partial class Channel
+    {
+        public async Task AddMembers(IEnumerable<string> userIDs, MessageInput msg = null)
+        {
+            var payload = new JObject(new JProperty("add_members", userIDs));
+            if (msg != null)
+            {
+                if (msg.User != null)
+                {
+                    msg.User = new User()
+                    {
+                        ID = msg.User.ID
+                    };
+                }
+                payload.Add("message", msg.ToJObject());
+            }
+
+            var request = this._client.BuildAppRequest(this.Endpoint, HttpMethod.POST);
+            request.SetJsonBody(JsonConvert.SerializeObject(payload));
+
+            var response = await this._client.MakeRequest(request);
+            if (response.StatusCode != System.Net.HttpStatusCode.Created)
+                throw StreamChatException.FromResponse(response);
+        }
+
+        public async Task RemoveMembers(IEnumerable<string> userIDs, MessageInput msg = null)
+        {
+            var payload = new JObject(new JProperty("remove_members", userIDs));
+            if (msg != null)
+            {
+                if (msg.User != null)
+                {
+                    msg.User = new User()
+                    {
+                        ID = msg.User.ID
+                    };
+                }
+                payload.Add("message", msg.ToJObject());
+            }
+
+            var request = this._client.BuildAppRequest(this.Endpoint, HttpMethod.POST);
+            request.SetJsonBody(JsonConvert.SerializeObject(payload));
+
+            var response = await this._client.MakeRequest(request);
+            if (response.StatusCode != System.Net.HttpStatusCode.Created)
+                throw StreamChatException.FromResponse(response);
+        }
+
+    }
+}
