+++ src/stream-chat-net/ExportChannel.cs
@@ -19,8 +19,6 @@ namespace StreamChat
         [JsonProperty(NullValueHandling = NullValueHandling.Ignore, PropertyName = "messages_until")]
         public DateTime? MessagesUntil { get; set; }
 
-        public ExportChannelRequest() { }
-
         public ExportChannelRequest WithChannelType(string type)
         {
             this.Type = type;
@@ -62,8 +60,6 @@ namespace StreamChat
 
         [JsonProperty(NullValueHandling = NullValueHandling.Ignore, PropertyName = "updated_at")]
         public DateTime? UpdatedAt { get; set; }
-
-        public ExportChannelsStatusResponse() { }
     }
 
     public class ExportChannelsResult
@@ -76,7 +72,5 @@ namespace StreamChat
 
         [JsonProperty(NullValueHandling = NullValueHandling.Ignore, PropertyName = "url")]
         public string URL { get; set; }
-
-        public ExportChannelsResult() { }
     }
 }
